requirements.txt
streamlit
pandas
yfinance
plotly
finta
python-dotenv
langchain-google-genai
langchain-core 